Split up the single file for testing.
Times are really hard and it will be hard to tell if this program is
working without good tests. `main.rs` files Don't seem suited for it, so
I split the program into a `lib.rs` and a `bin.rs`.

This commit also makes it so that the program filters times "correctly"
(adding quotes because I am not 100% sure it's working).