Generate the HTML in rust.
Last night I tried to get this to work using a python markdown parser.
This was nice because it meant that the executable didn't need to
generate it's own html. In the end this made it a little harder to
iterate and validate that the outputted text file was valid markdown.

This commit also updates the schedule to run much more frequently. The
plan is to undo this after it's clear that things are working.